+++ flutter_appauth_platform_interface/test/method_channel_flutter_appauth_test.dart
@@ -27,7 +27,7 @@ void main() {
     expect(
       log,
       <Matcher>[
-        isMethodCall('authorize', arguments: <String, Object>{
+        isMethodCall('authorize', arguments: <String, Object?>{
           'clientId': 'someClientId',
           'issuer': null,
           'redirectUrl': 'someRedirectUrl',
@@ -39,6 +39,7 @@ void main() {
           'allowInsecureConnections': false,
           'preferEphemeralSession': false,
           'promptValues': null,
+          'responseMode': null
         })
       ],
     );
@@ -47,11 +48,11 @@ void main() {
   test('authorizeAndExchangeCode', () async {
     await flutterAppAuth.authorizeAndExchangeCode(AuthorizationTokenRequest(
         'someClientId', 'someRedirectUrl',
-        discoveryUrl: 'someDiscoveryUrl', loginHint: 'someLoginHint'));
+        discoveryUrl: 'someDiscoveryUrl', loginHint: 'someLoginHint', responseMode: 'fragment'));
     expect(
       log,
       <Matcher>[
-        isMethodCall('authorizeAndExchangeCode', arguments: <String, Object>{
+        isMethodCall('authorizeAndExchangeCode', arguments: <String, Object?>{
           'clientId': 'someClientId',
           'issuer': null,
           'redirectUrl': 'someRedirectUrl',
@@ -67,7 +68,8 @@ void main() {
           'refreshToken': null,
           'authorizationCode': null,
           'grantType': 'authorization_code',
-          'codeVerifier': null
+          'codeVerifier': null,
+          'responseMode': 'fragment'
         })
       ],
     );
@@ -87,7 +89,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',
@@ -113,7 +115,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',
@@ -138,7 +140,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',