+++ flutter_appauth/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.1.0+2
+
+* Bumped AppAuth Android and iOS dependencies
+
+## 1.1.0+1
+
+* Updated the minimum version iOS supported by the example app to 9.0 since Flutter 2.5 has deprecated support for 8.0
+
+## 1.1.0
+
+* * Added the ability to specify the response mode for authorization requests. This can be done using the `responseMode` parameter  when constructing either an `AuthorizationRequest` or `AuthorizationTokenRequest`. This was done as the AppAuth Android SDK throws an exception when this was done via `additionalParameters`
+* Updated Dart SDK constraints
+
+## 1.0.0+1
+
+* There are no functional changes in this release. The only changes done were to suppress warnings that were occurring as a result of making use of Android v1 embedding APIs for backwards compatibility
+
+## 1.0.0
+
+* Migrated to null safety
+
 ## 0.9.2+6
 
 * [Android] community has reported that there seem to be instances where the plugin encounters a null intent on some devices upon processing a authorisation request. This resulted in a crash before but will now throw a `PlatformException`. Thanks to the PR from [Leon Havenga](https://github.com/li0nza)